separated initializing davis and davisio, because for the most readers of the code in davisio don't need to read initialize part
