+++ src/evl/core/initialize_davis.hpp
@@ -1,0 +1,9 @@
+// Copyright 2018 Event Vision Library.
+#ifndef SRC_EVL_CORE_INITIALIZE_DAVIS_HPP_
+#define SRC_EVL_CORE_INITIALIZE_DAVIS_HPP_
+
+#include "types.hpp"
+
+Davis initializeDavis(void);
+
+#endif  // SRC_EVL_CORE_INITIALIZE_DAVIS_HPP_
